==> src/notebooks/0.0-introduction.ipynb <==
<div style="text-align: center; background-color: #0A6EBD; font-family: 'Trebuchet MS', Arial, sans-serif; color: white; padding: 20px; font-size: 40px; font-weight: bold; border-radius: 0 0 0 0; box-shadow: 0px 6px 8px rgba(0, 0, 0, 0.2);">
    FIT-HCMUS, VNU-HCM 
    <br>
    Introduction To Data Science 
    <br>
    Final project 📌
</div>

<div style="text-align: center; background-color: #5A96E3; font-family: 'Trebuchet MS', Arial, sans-serif; color: white; padding: 20px; font-size: 40px; font-weight: bold; border-radius: 0 0 0 0; box-shadow: 0px 6px 8px rgba(0, 0, 0, 0.2);">
  Stage 00 - Project Introduction 📌
</div>

## Member informations

| Student ID | Name                |
| ---------- | ------------------- |
| 21127107   | Bui Vu The Minh     |
| 21127693   | Huynh Duc Thien     |
| 21127700   | Le Phuoc Thinh Tien |
| 21127704   | Pham Khanh Toan     |

## About the project

Our project is designed to showcase proficiency in the entire data science workflow. We start with the data collection phase, where we demonstrate the ability to crawl and gather relevant data from various sources. This is followed by preprocessing stage, and an in-depth exploratory data analysis (EDA) that uncovers patterns, trends, and insights within the dataset. Finally, we employ modeling techniques to make predictions or draw conclusions based on our analysis.

<br>

**Topic**
<br>
*Developing a Predictive Model for Football Player Performance* </br>
This data-AI project focuses on developing a predictive model for professional football player performance. By considering factors like fitness levels, historical stats, injuries, and situational context, the model aims to revolutionize player recruitment, providing a more efficient and data-driven approach for teams. The project's success could optimize strategic decision-making for coaches, leading to adaptive game plans and improved player development through personalized training programs.


## Main Stages

- **Stage 1**: Collecting data
- **Stage 2**: Prepressing data
- **Stage 3**: Explore data
- **Stage 4**: Data modelling
