==> src/notebooks/2.0-preprocessing.ipynb <==
<div style="text-align: center; background-color: #0A6EBD; font-family: 'Trebuchet MS', Arial, sans-serif; color: white; padding: 20px; font-size: 40px; font-weight: bold; border-radius: 0 0 0 0; box-shadow: 0px 6px 8px rgba(0, 0, 0, 0.2);">
    FIT-HCMUS, VNU-HCM 
    <br>
    Introduction To Data Science 
    <br>
    Final project 📌
</div>

<div style="text-align: center; background-color: #5A96E3; font-family: 'Trebuchet MS', Arial, sans-serif; color: white; padding: 20px; font-size: 40px; font-weight: bold; border-radius: 0 0 0 0; box-shadow: 0px 6px 8px rgba(0, 0, 0, 0.2);">
  Stage 02 - Preprocessing 📌
</div>

## **1. Set up enviroment**

# !pip install unidecode
# !pip install fuzzywuzzy
import pandas as pd
import numpy as np
import re
import warnings
from unidecode import unidecode
from fuzzywuzzy import fuzz
from fuzzywuzzy import process

warnings.filterwarnings("ignore")
pd.set_option('display.max_columns', None)

## **2. Reading and cleaning data**

### **2.1. Fbref data**

#### **Raw data inspect**

raw_data_fbref = pd.read_csv("../../data/raw/raw_data_fbref.csv")
raw_data_fbref.head()

**Checking data shape**

fbref_shape = raw_data_fbref.shape

# testing if data include enough 1000 items
print(f"Current shape: {fbref_shape}")

if fbref_shape[0] > 1000:
    print(f"Data is good!")
else:
    print(f"Raw data is absolutely small. Please recheck!")

**Checking duplicated rows**

index = raw_data_fbref.index
detect_dup_series = index.duplicated(keep='first')
num_duplicated_rows = detect_dup_series.sum()

# checking containing duplicated rows
if num_duplicated_rows == 0:
    print(f"Raw data have no duplicated rows.!")
else:
    if num_duplicated_rows > 1:
        ext = "lines"
    else:
        ext = "line"
    print(f"Raw data have {num_duplicated_rows} duplicated " + ext + ". Please de-deduplicate the data.!")

**Column information**

raw_data_fbref.info()

It seems there might be an issue with displaying the column names correctly due to duplicates. We'll modify the column names to ensure they present suitable meanings.

new_cols_name = ['Name', 'Position', 'PreferredFoot', 'Season', 'Age', 'Squad', 'Country', 
                 'Comp', 'LgRank', 'MP', 'Starts', 'Min', '90s', 'Gls', 'Ast', 'G+A', 'G-PK', 'PK', 
                 'PKatt', 'CrdY', 'CrdR', 'xG', 'npxG', 'xAG', 'npxG+xAG', 'PrgC', 'PrgP', 'PrgR',
                 'Gls/90', 'Ast/90', 'G+A/90', 'G-PK/90', 'G+A-PK/90', 'xG/90', 
                 'xAG/90', 'xG+xAG/90', 'npxG/90', 'npxG+xAG/90', 'Matches']

column_name_mapping = dict(zip(raw_data_fbref.columns, new_cols_name))
raw_data_fbref.rename(columns=column_name_mapping, inplace=True)

raw_data_fbref.columns

**Field meanings**

| Field            | Description|
|------------------|---------|
| **Name**         | The name ò the player|
| **Season**       | Season|
| **Age**          | Age at season start|
| **Squad**        | The team the player is playing for|
| **Position**     | Position on the field|
| **PreferredFoot**| Foot that the player prefer using|
| **Country**      | Name of the country|
| **Comp**         | Competition|
| **LgRank**       | Squad finish in competition|
| **MP**           | Matches Played|
| **Starts**       | Games that player in starting lineup|
| **Min**          | Minutes played|
| **90s**          | 90s Played (Minutes played divided by 90)|
| **Gls**          | Goals scored or allowed|
| **Ast**          | Assists|
| **G+A**          | Goals + Assists|
| **G-PK**         | Non-Penalty Goals|
| **PK**           | Penalty Kicks Made|
| **PKatt**        | Penalty Kicks Attempted|
| **CrdY**         | Yellow Cards|
| **CrdR**         | Red Cards|
| **xG**           | Expected Goals (xG totals include penalty kicks, excluding penalty shootouts)|
| **npxG**         | Non-Penalty Expected Goals|
| **xAG**          | Expected Assisted Goals|
| **npxG+xAG**     | Non-Penalty Expected Goals plus Assisted Goals|
| **PrgC**         | Progressive Carries (Carries that move the ball towards the opponent's goal line)|
| **PrgP**         | Progressive Passes (Completed passes that move the ball towards the opponent's goal line)|
| **PrgR**         | Progressive Passes Received (Completed passes that move the ball towards the opponent's goal line)|
| **Gls/90**       | Goals per 90 minutes (Minimum 30 minutes played per squad game to qualify as a leader)|
| **Ast/90**       | Assists per 90 minutes (Minimum 30 minutes played per squad game to qualify as a leader)|
| **G+A/90**       | Goals and Assists per 90 minutes (Minimum 30 minutes played per squad game to qualify as a leader)|
| **G-PK/90**      | Non-Penalty Goals per 90 minutes (Minimum 30 minutes played per squad game to qualify as a leader)|
| **G+A-PK/90**    | Non-Penalty Goals plus Assists per 90 minutes (Minimum 30 minutes played per squad game to qualify as a leader)|
| **xG/90**        | Expected Goals per 90 minutes (Minimum 30 minutes played per squad game to qualify as a leader)|
| **xAG/90**       | Expected Assisted Goals per 90 minutes (Minimum 30 minutes played per squad game to qualify as a leader)|
| **xG+xAG/90**    | Expected Goals plus Assisted Goals per 90 minutes (Minimum 30 minutes played per squad game to qualify as a leader)|
| **npxG/90**      | Non-Penalty Expected Goals per 90 minutes (Minimum 30 minutes played per squad game to qualify as a leader)|
| **npxG+xAG/90**  | Non-Penalty Expected Goals plus Assisted Goals per 90 minutes (Minimum 30 minutes played per squad game)|

**Removing meaningless rows**

- In certain football leagues, players may miss a season due to various reasons such as injuries, non-participation, suspensions, transfers, and more. This results in rows with incomplete information in the dataset. Therefore, we are contemplating the removal of these rows.

raw_data_fbref.dropna(subset=['Season'], inplace=True)
raw_data_fbref.shape

- Some players were registered but did not attempt any games due to some specific reasons (not suitable for tactic,...), rendering their data irrelevant. It is necessary to delete these rows.

raw_data_fbref = raw_data_fbref[raw_data_fbref['MP'] != 0]
raw_data_fbref.shape

- Goalkeepers are rarely injured, so we can drop them as collecting data in this group will not yield the demand proportion

raw_data_fbref = raw_data_fbref[raw_data_fbref['Position'] != ' GK']
raw_data_fbref.shape

**Remove unnecessary columns**

raw_data_fbref['Matches'].unique()

Column `Matches` only has Matches value. In fact, on website this is a hyperlink to another match so dropping this column is necessary

raw_data_fbref = raw_data_fbref.drop(['Matches'],axis = 1)
raw_data_fbref.columns

**Reformating features and Adjusting data types**

- Overall, data type of most fields is correct except for minutes played `Min`, which should be integer instead of object. Here we're going to remove the comma in each value.

def format_min(min):
     if pd.notna(min):
        return int(min.replace(',', ''))
     else:
        return min
    
raw_data_fbref['Min'] = raw_data_fbref['Min'].apply(format_min)

- Additional players personal infos, and fundamental statistics such as `MP`, `Gls`, `Ast`, `G+A`, etc., need to be converted to integers. However, a challenge arises because these columns contain NaN values, which are represented as floating-point numbers, often indicating missing data. To facilitate the conversion of these columns to integers, the transformation from NaN to zero must be performed prior to converting the remaining values.

int_cols = ['Age', 'MP', 'Starts', 'Min', 'Gls', 'Ast', 'G+A', 
            'G-PK', 'PK', 'PKatt', 'CrdY', 'CrdR', 
            'PrgC', 'PrgP', 'PrgR']

raw_data_fbref[int_cols] = raw_data_fbref[int_cols].fillna(0).astype(int)
raw_data_fbref.dtypes

- Extract the the teams' places in `LgRank` columns.

raw_data_fbref['LgRank'] = raw_data_fbref['LgRank'].str.extract('(\d+)').astype(int)
raw_data_fbref['LgRank'].unique()

- Remove redundant letters in values of `Country` column, which in fact is the national flags, causing by crawling process.

# Countries names format and their frequency
count = raw_data_fbref["Country"].value_counts()
count

# get rid of the first word of countries name
raw_data_fbref['Country'] = raw_data_fbref['Country'].str.split().str[1]
count = raw_data_fbref["Country"].value_counts()
count

- Data of players currently playing in premier league may include their previous records from other leagues in which `Season` format is different. It is neccessary to adjust this for later stages.

season_count = raw_data_fbref['Season'].value_counts()
print(season_count)
print(f"total: {season_count.sum()}")

def format_season(season):
    if pd.notna(season) and '-' in season:
        return season
    elif pd.notna(season):
        return f"{season}-{int(season) + 1}"
    else:
        return season
raw_data_fbref['Season'] = raw_data_fbref['Season'].apply(format_season)

season_count = raw_data_fbref['Season'].value_counts()
print(season_count)
print(f"total: {season_count.sum()}")

- Remove prefix number, and normalize league name in `Comp` column using `unicode` library.

raw_data_fbref['Comp'] = raw_data_fbref['Comp'].apply(unidecode)
raw_data_fbref['Comp'] = raw_data_fbref['Comp'].str.replace(r'^(\d+\.?\s*)+', '', regex=True)
comp_counts = raw_data_fbref['Comp'].value_counts()
comp_counts

- Each player may hold multiple positions on the field based on the coach's tactics. However, for the sake of simplicity in this project, we will focus on their main position.

position_count = raw_data_fbref['Position'].value_counts()
position_count

We will take the first (main) position of each player.

raw_data_fbref['Position'] = raw_data_fbref['Position'].str.replace('-', ' ').str.split().str[0]
position_count = raw_data_fbref['Position'].value_counts()
position_count

**Handling missing values**

First we will have a look at the missing ratio of each columns:

missing_ratio = raw_data_fbref.isnull().mean() * 100
missing_ratio

The missing values in the columns can be categorized into two groups. The first group consists of columns with a low missing ratio, even less than 1%. The second group includes columns with a substantial number of missing values, approximately around 30%.

- First step, we consider removing the rows that contains more than 50% NaN values.

raw_data_fbref = raw_data_fbref.dropna(thresh=(raw_data_fbref.shape[1] * 0.5))
raw_data_fbref.shape

- Despite certain columns having a high missing ratio, we opt to retain them for future analysis given their significance. As this pertains to football statistics, a more equitable approach involves filling the empty cells with mean values derived from players occupying similar positions on the field.

missing_cols = missing_ratio[missing_ratio > 0].index
raw_data_fbref[missing_cols] = raw_data_fbref.groupby('Position')[missing_cols].transform(lambda x: x.fillna(x.mean()))

missing_ratio = raw_data_fbref.isnull().mean() * 100
if (missing_ratio == 0).all():
    print("The data is now good!")
else:
    print("Still having missing values, please recheck!")

**Processed data**

processed_data_fbref = raw_data_fbref
processed_data_fbref.head()

### **2.2. Transfermarkt data**

#### **Raw data inspect**

raw_data_trans = pd.read_csv("../../data/raw/raw_data_trans.csv")
raw_data_trans.head()

It seems that, for this data table, the data is relatively clean, and there are no columns containing junk data. Therefore, removing columns is not yet necessary at the current time. However, some reformation must be done for synchronization with the previous data.

**Checking data shape**

trans_shape = raw_data_trans.shape
trans_shape

# testing if data include enough 1000 items
print(f"Current shape: {fbref_shape}")

if fbref_shape[0] > 1000:
    print(f"Data is good!")
else:
    print(f"Raw data is absolutely small. Please recheck!")

**Checking duplicated rows**

index = raw_data_trans.index
detect_dup_series = index.duplicated(keep='first')
num_duplicated_rows = detect_dup_series.sum()

# checking containing duplicated rows
if num_duplicated_rows == 0:
    print(f"Raw data have no duplicated rows.!")
else:
    if num_duplicated_rows > 1:
        ext = "lines"
    else:
        ext = "line"
    print(f"Raw data have {num_duplicated_rows} duplicated " + ext + ". Please de-deduplicate the data.!")

**Column informations**

raw_data_trans.info()

The dataset appears to be well-maintained with no discernible issues, and there is no necessity to eliminate any columns or rows.

**Field meanings**

| Field           | Description                                                  |
|-----------------|--------------------------------------------------------------|
| **Name**        | The name of the player.                                      |
| **Season**      | The season during which the injury occurred.                 |
| **Injury**      | The type or nature of the injury sustained by the player.    |
| **From**        | The date when the player suffered the injury.                |
| **Until**       | The date when the player recovered from the injury.          |
| **Days**        | The duration of the injury in days.                          |
| **Games missed**| The number of games the player missed due to the injury.     |


**Reformating features and Adjusting data types**

- We need to change `Season` format to match the one of fbref data for later stage.

seas_count = raw_data_trans['Season'].value_counts()
print(seas_count)
print(f"Total: {seas_count.sum()}")

def format_trans_season (season):
    if pd.notna(season):
        year = season.split('/')
        return f"20{year[0]}-20{year[1]}"
    else:
        return season
raw_data_trans['Season'] = raw_data_trans['Season'].apply(format_trans_season)
seas_count = raw_data_trans['Season'].value_counts()
print(seas_count)
print(f"Total: {seas_count.sum()}")

- Adjusting `Days` format to int, and removing unnecessary 'days' string.

raw_data_trans  = raw_data_trans.copy()
raw_data_trans['Days'] = raw_data_trans['Days'].str.split().str[0].astype('int64')
raw_data_trans['Days'].dtypes

- We will adjust the date object into simpler format.

raw_data_trans['from'] = pd.to_datetime(raw_data_trans['from'], errors='coerce')
raw_data_trans['until'] = pd.to_datetime(raw_data_trans['until'], errors='coerce')
raw_data_trans[['from', 'until']].head()

- The format of the `Games missed` column appears to be incorrect. Let's examine its values first.

missed_games_count = raw_data_trans['Games missed'].value_counts()
missed_games_count

It appears that the data uses some characters to represent zero games missed. To address this, we should replace these characters with 0 and then convert the column to the correct data type.

raw_data_trans['Games missed'] = pd.to_numeric(raw_data_trans['Games missed'].replace('[^\d]', '0', regex=True))
raw_data_trans['Games missed']

**Handling missing values**

First we will have a look at the missing ratio of each columns:

missing_ratio = raw_data_trans.isnull().mean() * 100
missing_ratio

Only one column exhibits missing values, possibly as a result of the reformation process. Given the relatively small proportion and the potential hindrance in subsequent data processing, we have decided to remove the rows with missing values in this column.

raw_data_trans = raw_data_trans.dropna(subset=['until'])

missing_ratio = raw_data_trans.isnull().mean() * 100
if (missing_ratio == 0).all():
    print("The data is now good!")
else:
    print("Still having missing values, please recheck!")

**Processed data**

processed_data_trans = raw_data_trans
processed_data_trans.head()

## 3. Merging data

**Merging 2 dataframes can help to attach additional informations for later stage**

There is a challenge where the names of players in two dataframes may not always match. Therefore, we require the assistance of a package called `fuzzywuzzy` to assess the similarity between two strings for merging purposes.

# create a dictionary for matching names of 2 Name columns of 2 dataframes
name_matches = {}
fbref_name = processed_data_fbref['Name'].unique()
trans_name = list(processed_data_trans['Name'].unique())

for name in fbref_name:
    if trans_name:
        best_match, score = process.extractOne(name, trans_name,scorer = fuzz.token_set_ratio)
        if score > 90:
            name_matches[name] = best_match
            trans_name.remove(best_match)
        else:
            name_matches[name] = name
    else:
        name_matches[name] = name


processed_data_fbref['Name'] = processed_data_fbref['Name'].map(name_matches)

# Merge based on both 'Name' and 'Season'
merged_df = pd.merge(processed_data_fbref, processed_data_trans , left_on=['Name', 'Season'], right_on=['Name', 'Season'], how='left')

# matching names found
pd.Series(name_matches)

merged_df

**Check if merging is correct**

assert len(processed_data_fbref['Name'].unique()) == len(name_matches.keys())
assert len(merged_df['Season'].unique()) == len(processed_data_fbref['Season'].unique())

## 4. Save processed data

# save data to processed data folder
merged_df.to_csv('../../data/processed/processed_fb_data.csv', index=False, encoding='utf-8')
