==> src/notebooks/4.0-feature-encoding-scaling.ipynb <==
<div style="text-align: center; background-color: #0A6EBD; font-family: 'Trebuchet MS', Arial, sans-serif; color: white; padding: 20px; font-size: 40px; font-weight: bold; border-radius: 0 0 0 0; box-shadow: 0px 6px 8px rgba(0, 0, 0, 0.2);">
    FIT-HCMUS, VNU-HCM 
    <br>
    Introduction To Data Science 
    <br>
    Final project 📌
</div>

<div style="text-align: center; background-color: #5A96E3; font-family: 'Trebuchet MS', Arial, sans-serif; color: white; padding: 20px; font-size: 40px; font-weight: bold; border-radius: 0 0 0 0; box-shadow: 0px 6px 8px rgba(0, 0, 0, 0.2);">
  Stage 04 - Feature encoding and scaling 📌
</div>

# Import library

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
from scipy.stats import skew
from sklearn.preprocessing import StandardScaler,RobustScaler

df = pd.read_csv('../../data/processed/processed_fb_data.csv')

df.head()

# Preprocessing 

The reason to drop these columns : 
- The Squad column don't provide much information bacause the dataset has LgRank whose meaning is the rank of the squad in the league . The teams in the top of table usually have better stats than the teams in the opposite side.
- The Country,Comp column just show the name of country,league of the player's club . Although each country can have different style of playing football such as Brazil is famous for attacking, Italy is known as good defending ,...   ,it is not enough information to judge all countries ,leagues
- These 2 columns : from , until show the begining and ending of the injury . The dataset has more informative columns is Days (counting the number of days player missed)


df.drop(['Squad','Country','Comp','from','until'],inplace=True,axis = 'columns')

#Re-define the meaning of Injury column : Injury column is used to identify that the player has the injury or not
df['Injury'] = df['Injury'].apply(lambda x : 0 if x is np.nan else 1)

#Fill missing values
df['Days'].fillna(0,inplace=True)
df['Games missed'].fillna(0,inplace=True)

Calculate all performance stats and injury stats according to each season of each player

#create a dictionary for each column in performance and injury stats with a function aggregate 
agg_dict = {}
for col in df.columns[6:]:
    if '90' in col:
        agg_dict[col]='mean'
    else:
        agg_dict[col] = 'sum'

info_df = df.groupby(['Name','Season']).agg(agg_dict)
info_df.head()

df.drop(df.columns[6:],axis='columns',inplace=True)

#merge the info_df to the main df by 2 column name and season
df = df.merge(right=info_df,how='inner',on=['Name','Season'])

#drop duplicate column because the df has some duplicate name and season when player play in 2 different team in 1 season
df.drop_duplicates(keep='first',inplace=True)

After being used to calculate the performance and injury stats, the Name and Season columns become unnecessary so we decide to drop these column

df.drop(['Name','Season'],axis='columns',inplace=True)

df.head()

cate_col = df.select_dtypes(include=object).columns
num_col = df.select_dtypes(exclude=object).columns

# One-hot encoding for categorical column

#Find categorical column in the dataframe
cate_col

#One-hot encoding for cate_col
df = pd.get_dummies(df, columns=cate_col)
df

#convert from bool type to int type 
bool_col = df.select_dtypes(include=bool)
for col in bool_col:
    df[col] =df[col].astype(np.int64)

df.reset_index(inplace=True,drop=True)

df

df.to_csv("../../data/processed/encoded_data.csv", index=False)

# Feature scaling for numerical columns

#Find numerical columns
num_col

Calculate the percentage of outliers in each column

def count_outlier(df,num_col):
    outlier_dict ={}
    for col in num_col:
        Q1 = np.percentile(df[col],25)
        Q3 = np.percentile(df[col],75)
        iqr = Q3 - Q1
        outliers = ((df[col] < (Q1 - 1.5 * iqr)) | (df[col] > (Q3 + 1.5 * iqr)))
        outlier_dict[col] = outliers.sum() / len(df[col])
    return outlier_dict

outlier_dict =count_outlier(df,num_col)
outlier_df = pd.DataFrame(list(outlier_dict.items()), columns=['Column', 'Outlier Percentage'])
outlier_df['Skew'] = [skew(df[col],bias=False) for col in num_col]
outlier_df.sort_values('Outlier Percentage',ascending=False)

With the column having high outliers percentages, we apply RobustScaler to scale the data because the large number of outliers can affect the dataset and the accuracy of model in the next section

# find the column with high percentage of outliers ( Percentage >= 5%)
highoutlier_col = outlier_df.loc[outlier_df['Outlier Percentage']>=0.05]['Column'].to_list()
print(highoutlier_col)

#Apply RobustScaler for these columns
robustscaler = RobustScaler()
robustscaler.fit(df[highoutlier_col])
scaled_robust_data = robustscaler.transform(df[highoutlier_col])
df[highoutlier_col] = scaled_robust_data

With the column having positive skewness ,we apply Cube root transformation to reduce the skewness of data . THe reason why we don't choose Square Root transformation is that the data in many columns has zero 

#find the column with positive skewness and don't choose columns appearing in highoutlier_col
pos_skew_col = outlier_df.loc[outlier_df['Skew']>=outlier_df['Skew'].median()]['Column'].to_list()
pos_skew_col = [x for x in pos_skew_col if x not in highoutlier_col]
print(pos_skew_col)

#Apply Cube Root Transformation for these columns
for col in pos_skew_col:
    df[col] = df[col].pow(1/3)

For the remaining columns , except 2 column LgRank and 90s don't have outliers, we apply StandardScaler to these columns because these columns don't have many outliers and the skew of these columns is not too high

remain_col = [x for x in outlier_df['Column'] if x not in pos_skew_col and x not in highoutlier_col ]
remain_col.remove('LgRank')
remain_col.remove('90s')
remain_col

stdscaler = StandardScaler()
stdscaler.fit(df[remain_col])
scaled_std_data = stdscaler.transform(df[remain_col])
df[remain_col] = scaled_std_data

# Save to csv file

df.to_csv("../../dat a/processed/scaled_encoded_data.csv", index=False)
